requirements.txt
phidata
openai
duckduckgo-search
fastapi[standard]
sqlalchemy
lancedb
tantivy
pypdf
python-dotenv
pandas
gradio